+++ crates/ryo3-jiff/src/ry_time.rs
@@ -123,13 +123,7 @@ impl RyTime {
     }
 
     fn __repr__(&self) -> String {
-        format!(
-            "Time(hour={}, minute={}, second={}, nanosecond={})",
-            self.0.hour(),
-            self.0.minute(),
-            self.0.second(),
-            self.0.nanosecond()
-        )
+        format!("{self}")
     }
 
     fn isoformat(&self) -> String {
@@ -445,7 +439,14 @@ impl RyTime {
 
 impl Display for RyTime {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "Time<{}>", self.0)
+        write!(
+            f,
+            "Time(hour={}, minute={}, second={}, nanosecond={})",
+            self.0.hour(),
+            self.0.minute(),
+            self.0.second(),
+            self.0.nanosecond()
+        )
     }
 }
 impl From<Time> for RyTime {